% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MarketingAnalytics_functions.R
\name{calculate_segment_means}
\alias{calculate_segment_means}
\title{Calculate Segment Means}
\usage{
calculate_segment_means(df)
}
\arguments{
\item{x}{The number you wish to add commas to.}
}
\description{
This function allows you to group every three digits in a number using a comma.
}
\examples{


}
\keyword{comma}
