% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MarketingAnalytics_functions.R
\name{Estimate_Market_Shares_old}
\alias{Estimate_Market_Shares_old}
\title{Estimate market share old version}
\usage{
Estimate_Market_Shares_old(product_alternatives, pws)
}
\arguments{
\item{product_alternatives, pws}{product alternatives and partworths table}
}
\description{
This function estimates market shares for new product alternatives
}
\keyword{market}
\keyword{shares}
