% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MarketingAnalytics_functions.R
\name{Estimate_Partworth_Utilities}
\alias{Estimate_Partworth_Utilities}
\title{Estimate partworth utilities}
\usage{
Estimate_Partworth_Utilities(ratings, products)
}
\arguments{
\item{ratings, products}{customer ratings of products and description table of products}
}
\description{
This function estimates partworth utilities
}
\keyword{partworth}
\keyword{utilities}
