% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MarketingAnalytics_functions.R
\name{Find_Optimal_Products}
\alias{Find_Optimal_Products}
\title{Find optimal products}
\usage{
Find_Optimal_Products(MS, rule, n)
}
\arguments{
\item{MS}{market shares}

\item{rule}{decision rule}

\item{n}{n is Max_No_of_products_to_return}
}
\description{
This function finds optimal products
}
\keyword{optimal}
\keyword{products}
