% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MarketingAnalytics_functions.R
\name{is.defined}
\alias{is.defined}
\title{Check if object is defined in the environment}
\usage{
is.defined(thisobject)
}
\arguments{
\item{thisobject}{name of the object}
}
\description{
Check if object is defined in the environment
}
\keyword{defined}
\keyword{object}
