% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MarketingAnalytics_functions.R
\name{is.sig}
\alias{is.sig}
\title{Check for significant difference}
\usage{
is.sig(x, y)
}
\arguments{
\item{x, y}{x and y vectors to be compared with a two-sided T-test}
}
\description{
This function checks for significant difference in means of two vectors
}
\examples{

}
\keyword{averages}
\keyword{calculate}
