% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MarketingAnalytics_functions.R
\name{Plot_MS}
\alias{Plot_MS}
\title{Plot market shares}
\usage{
Plot_MS(MS, rule)
}
\arguments{
\item{MS, mytext}{Market share and message}
}
\description{
This function plots market shares
}
\keyword{market}
\keyword{plot}
\keyword{shares}
