% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MarketingAnalytics_functions.R
\name{Plot_MS_rule1}
\alias{Plot_MS_rule1}
\title{Plot market shares for rule 1}
\usage{
Plot_MS_rule1(MS, mytext = "")
}
\arguments{
\item{MS}{Market share}

\item{mytext}{message}
}
\description{
This function plots market shares
}
\keyword{market}
\keyword{plot}
\keyword{shares}
