% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MarketingAnalytics_functions.R
\name{Plot_MS_rule3}
\alias{Plot_MS_rule3}
\title{Plot market shares for rule 3}
\usage{
Plot_MS_rule3(MS, mytext = "")
}
\arguments{
\item{MS, mytext}{Market share and message}
}
\description{
This function plots market shares
}
\keyword{market}
\keyword{plot}
\keyword{shares}
