% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MarketingAnalytics_functions.R
\name{is.sig}
\alias{is.sig}
\title{Check for significant difference}
\usage{
is.sig(x, y)
}
\arguments{
\item{x}{vector to be compared with y in a two-sided T-test}

\item{y}{vector to be compared with x in a two-sided T-test}
}
\value{
1 if significant difference in means, 0 if no sig. difference in means
}
\description{
This function checks for significant difference in means of two vectors
}
\examples{
x <- 1:20
y <- 20:30
is.sig(x,y)

}
\keyword{averages}
\keyword{calculate}
